refactor: drop the server-side OG cache — Cloudflare already is the cache
Measured over 48h of production traffic, the origin sees an almost perfectly
unique stream of OG requests: 1,189 across 1,134 distinct cards, 1,081 of them
requested exactly once, max 4 repeats. A *perfect* origin cache could have
served 4% of them. That is not a coincidence — these responses carry
`public, max-age=…` and Cloudflare serves the repeats from the edge (verified:
MISS then HIT on a real card), so by construction the origin only ever sees
misses.

We were paying for that 4% twice over. First with ocache's unbounded per-process
`Map` of webp bytes, which is one of the things that OOM-killed the workers.
Then, in this branch, with an `og_cache` KV table — a base64 round-trip
inflating every card 33% into SQLite TEXT, a sweep on the 15-min timer, two
gauges, and another writer to the file we already have to VACUUM. Rendering on
demand costs ~600ms p50 on a worker thread, ~25 times an hour.

So: no server-side cache. `src/utils/ogCache.ts` and its test are deleted, along
with the `og:` mount, the sweep, and `bookhive_og_cache_entries`/`_bytes`. That
also retires the import-cycle workaround the module existed to dodge — nothing
in `src/context.ts` reaches into the OG path any more.

What survives is `renderOnce` in `src/routes/og.tsx`: concurrent requests for
the *same* cold card share one render rather than starting N. It holds promises,
never bytes, and always clears in `finally` — the one failure mode the edge
cache genuinely cannot cover, and the same class of unbounded fan-out this
branch is about everywhere else.
